Fix CI gates: clippy nesting, SPEC_0021 file size, python construction
- Extract build_scalarized_children_index (and the new selection
  odometer) as free functions: clippy's excessive-nesting gate rejected
  the inline loops.
- Split the wgsl-solve render tests into codegen/wgsl_solve_tests.rs:
  the added tests pushed codegen_tests.rs past the SPEC_0021 2,000-line
  hard limit (now 1,985).
- CompilationResult gained a private cache field, breaking
  rumoca-bind-python's struct-literal construction; add
  CompilationResult::new and use it in both the CLI and the bindings.
